erik

+++ mysite/news/models.py
@@ -9,7 +9,7 @@ class News(models.Model):
     update_at = models.DateTimeField(auto_now_add=True, verbose_name='Обновлено')
     photo = models.ImageField(upload_to='photos/%Y/%m/%d', verbose_name='Фото', blank=True)
     is_published = models.BooleanField(default=True, verbose_name='Опубликовано')
-    views = models.IntegerField(default=0, verbose_name='Количество просмотров')
+    views = models.BigIntegerField(default=0, verbose_name='Количество просмотров')
     category = models.ForeignKey('Category', on_delete=models.PROTECT, null=True,
                                  verbose_name='Категория')
     author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='author', verbose_name='Автор', default=True)
